fix: droplet marked as STOPPED incorrectly
This patch fixes the bug when droplet.exited with reason "STOPPED" is
interpreted to mean that the application is stopped.

While it holds true in most of the cases, it does not hold true for
the case of downscaling an app to fewer instances. In this case, extra
instances are explicitly stopped by the cc.

The bug is subtle because the true application state is soon restored
using the bulk_api, so there's a short window of incorrect state,
based on which hm can make a wrong decision.

Change-Id: I3f0354bab1080f73cb1c73f25206df2071f079eb